+++ angular2-token.d.ts
@@ -1,2 +1,3 @@
 export * from './src/angular2-token.model';
 export * from './src/angular2-token.service';
+export * from './src/a2t-ui/a2t-ui.module';