+++ src/a2t-ui/a2t-shared/a2t-headline/a2t-headline.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'a2t-headline',
+    template: '<p><ng-content></ng-content></p>',
+    styleUrls: ['a2t-headline.component.css']
+})
+export class A2tHeadlineComponent {
+    constructor() { }
+}