+++ src/a2t-ui/a2t-ui.component.ts
@@ -1,0 +1,11 @@
+import { Component, Input } from '@angular/core';
+import { Angular2TokenService } from '../angular2-token.service';
+
+@Component({
+    selector: 'a2t-ui',
+    templateUrl: 'a2t-ui.component.html',
+    styleUrls: ['a2t-ui.component.css']
+})
+export class A2tUiComponent {
+    constructor() { }
+}